Modify expected result to match actual result better
This expected result change cannot be justified by an analytic solution
for what we are attempting to model.  However, it may be our modeling
which is deficient.  Perhaps a better interpretation of the velocity
definition is there is a +2.0 dipole at the center of our central voxel,
and a -1.0 dipole on the left/right neighbors.  This could explain why
the field decays faster than for a simple dipole.